LPS-90074 Move YAML parsing logic to portal-vulcan-api.
