+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/config/Application.java
@@ -12,39 +12,48 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.config;
+package com.liferay.portal.vulcan.yaml.config;
 
 /**
  * @author Peter Shin
  */
-public class Security {
+public class Application {
 
-	public String getBasicAuth() {
-		return _basicAuth;
+	public String getBaseURI() {
+		return _baseURI;
 	}
 
-	public String getGuestAllowed() {
-		return _guestAllowed;
+	public String getClassName() {
+		return _className;
 	}
 
-	public String getOAuth2() {
-		return _oAuth2;
+	public String getName() {
+		return _name;
 	}
 
-	public void setBasicAuth(String basicAuth) {
-		_basicAuth = basicAuth;
+	public Security getSecurity() {
+		return _security;
 	}
 
-	public void setGuestAllowed(String guestAllowed) {
-		_guestAllowed = guestAllowed;
+	public void setBaseURI(String baseURI) {
+		_baseURI = baseURI;
 	}
 
-	public void setOAuth2(String oAuth2) {
-		_oAuth2 = oAuth2;
+	public void setClassName(String className) {
+		_className = className;
 	}
 
-	private String _basicAuth;
-	private String _guestAllowed;
-	private String _oAuth2;
+	public void setName(String name) {
+		_name = name;
+	}
+
+	public void setSecurity(Security security) {
+		_security = security;
+	}
+
+	private String _baseURI;
+	private String _className;
+	private String _name;
+	private Security _security;
 
 }
