+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/openapi/Components.java
@@ -12,39 +12,32 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.openapi;
+package com.liferay.portal.vulcan.yaml.openapi;
+
+import java.util.Map;
 
 /**
  * @author Peter Shin
  */
-public class Items {
+public class Components {
 
-	public String getFormat() {
-		return _format;
+	public Map<String, Parameter> getParameters() {
+		return _parameters;
 	}
 
-	public String getReference() {
-		return _reference;
+	public Map<String, Schema> getSchemas() {
+		return _schemas;
 	}
 
-	public String getType() {
-		return _type;
+	public void setParameters(Map<String, Parameter> parameters) {
+		_parameters = parameters;
 	}
 
-	public void setFormat(String format) {
-		_format = format;
+	public void setSchemas(Map<String, Schema> schemas) {
+		_schemas = schemas;
 	}
 
-	public void setReference(String reference) {
-		_reference = reference;
-	}
-
-	public void setType(String type) {
-		_type = type;
-	}
-
-	private String _format;
-	private String _reference;
-	private String _type;
+	private Map<String, Parameter> _parameters;
+	private Map<String, Schema> _schemas;
 
 }
