+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/java/com/liferay/portal/vulcan/yaml/openapi/Delete.java
@@ -12,32 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.tools.rest.builder.internal.yaml.openapi;
-
-import java.util.Map;
+package com.liferay.portal.vulcan.yaml.openapi;
 
 /**
  * @author Peter Shin
  */
-public class Components {
-
-	public Map<String, Parameter> getParameters() {
-		return _parameters;
-	}
-
-	public Map<String, Schema> getSchemas() {
-		return _schemas;
-	}
-
-	public void setParameters(Map<String, Parameter> parameters) {
-		_parameters = parameters;
-	}
-
-	public void setSchemas(Map<String, Schema> schemas) {
-		_schemas = schemas;
-	}
-
-	private Map<String, Parameter> _parameters;
-	private Map<String, Schema> _schemas;
-
+public class Delete extends Operation {
 }
