+++ modules/apps/portal-vulcan/portal-vulcan-api/src/main/resources/com/liferay/portal/vulcan/yaml/config/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
